Add find method
The find method receives a value to search on the tree. First it checks
if the root value is equal to the value we're searching, it returns
this node if it is. Then it checks if the value we're looking for is
lower than the current node, if it is then we search in the left part
of the tree because there may be the value. If the value we're searching
is bigger than the current node then it search in the right part of the
tree. We do these until we found the node or we found a nil value.