Add level_order method
This method uses a queue to travel the tree by levels. It first add the
root node, then enters a loop until the queue is empty it removes the
first element of the queue and assigns it to a variable, then it adds
to the queue the left and right children if they exist. At the end it
returns the array of elements ordered by level.