Add blocks support to level_order method
Now the level_order method has support for blocks. It checks if a block
is given before yielding the values, also it checks the same and creates
the tree_elements array to create it when not block is given, appending
values and returning them at the end of the method.