Add inorder method
The inorder method represents an inorder traversal of a tree. It
consists in moving to the left subtree, visiting the root and moving to
the right subtree. It uses recursion to make this steps.

For the moment, it only stores the nodes's values in an array and return
it at the end of the method.