Add height method
This method calculates the height of a given node. To calculate this,
the method uses recursion and has two parameters: the current node and
the current height. First the base case if the node is nil or both
children are nil, then it returns the current height. Then it calculates
the height of both the left and right children recursevily and finally
it checks which one is heighest and returns that one.