Add balanced? method
This method verifies if the tree is balanced. A balanced tree is one
where the difference between its left and right subtrees heights are 1
or less.

We use recursion to travel the tree; if the node is nil, then return.
Then calculate the left subtree height and then the right subtree
height. After this, calculate the difference between these two heights.
If the difference is greater than or equal to 2, return false; if not
then check if the left and subtrees are balanced. After this return
true.