Add delete method
This method receives a value and uses the root node at the first call
because it travels the tree using recursion to find the node to delete.

For the moment it only deletes nodes when the node is a leaf node or
it's a node that has only one child. When it finds a leaf node, it
returns nil; and when it finds a node with one child it checks if it has
a left or right child, then it assigns the child node in a variable, set
the node to delete to nil and return the child node.